+++ src/main/java/dev/ftb/packcompanion/features/actionpad/ActionPadItem.java
@@ -1,6 +1,6 @@
-package dev.ftb.packcompanion.features.teleporter;
+package dev.ftb.packcompanion.features.actionpad;
 
-import dev.ftb.packcompanion.features.teleporter.net.OpenTeleporterPacket;
+import dev.ftb.packcompanion.features.actionpad.net.OpenActionPadPacket;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResultHolder;
@@ -10,15 +10,15 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 import net.neoforged.neoforge.network.PacketDistributor;
 
-public class TeleporterItem extends Item {
-    public TeleporterItem(Properties properties) {
+public class ActionPadItem extends Item {
+    public ActionPadItem(Properties properties) {
         super(properties);
     }
 
     @Override
     public InteractionResultHolder<ItemStack> use(Level level, Player player, InteractionHand usedHand) {
         if (!level.isClientSide()) {
-            PacketDistributor.sendToPlayer((ServerPlayer) player, new OpenTeleporterPacket(TeleporterDestinations.get().getUnlockedDestinations(player)));
+            PacketDistributor.sendToPlayer((ServerPlayer) player, new OpenActionPadPacket(PadActions.get().getUnlockedActions(player)));
         }
 
         return super.use(level, player, usedHand);