+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-## [21.1.7]
+## [21.1.10]
+
+### Changed
+* `Teleporter item` has been changed to `Action Pad` to better reflect it's more generic nature
+
+### Fixed
+* Shader notice should work again
+
+## [21.1.9]
 
 ### Added
 * Teleporting item